backend/app/utils/extract_text.py
import fitz  # pymupdf
import docx
import io


def extract_text(file_bytes: bytes, file_type: str) -> str:
    if file_type == "pdf":
        doc = fitz.open(stream=file_bytes, filetype="pdf")
        return "\n".join(page.get_text() for page in doc)

    elif file_type == "docx":
        doc = docx.Document(io.BytesIO(file_bytes))
        return "\n".join(p.text for p in doc.paragraphs if p.text.strip())

    elif file_type == "txt":
        return file_bytes.decode("utf-8", errors="ignore")

    raise ValueError(f"Unsupported file type: {file_type}")
