backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]>=3.1.0
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
anthropic>=0.85.0
boto3==1.35.28
python-dotenv==1.0.1
pydantic>=2.10.0
email-validator>=2.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
stripe>=10.0.0
